pyproject.toml
[project]
name = "smartparsal"
version = "0.1.0"
description = "SmartParcel"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.7",
    "django-crispy-forms>=2.4",
    "django-environ>=0.12.0",
    "django-phonenumber-field[phonenumbers]>=8.3.0",
    "djangorestframework>=3.16.1",
    "pre-commit>=4.3.0",
    "psycopg2-binary>=2.9.10",
]
[tool.ruff]
line-length = 88
extend-select = ["I"]
exclude = [
    "migrations",
    "venv",
    "env",
    "__pycache__",
]
